More user-friendly update
This new code allows the user to input the front and back of each card at the same time, which is more similar to the process of making notecards in real life.  This code also is more friendly to those with less programming experience, by changing some of the print statements to include less of the python language.  Overall, these changes just make the notecards more user friendly.